Back intake Subsystem (#8)
* BackIntake Subsystem created

* Back Intake Subsystem Revamp

used the front intake model to make the subsystem more consistent compared to the rest of the code, kept an isOn() feature not in the front intake Subsystem

* Fixed Constant Number and renamed some variabled

changed m_backIntakeMotor to BeltMotor and INTAKE_BACK to BELT to make it match the docs page

* Added Second Belt motor and Fixed constant number

The motor has is inverse to what the first belt motor is. For Constants BELT1 = 10 , BELT2 =11. (belt 2 is the new one)

* Simplified isOn() boolean code

* Name fixes for consistency

* Update BackIntakeSubsystem.java

---------

Co-authored-by: Carson Going <117052774+GarlicBreadPog@users.noreply.github.com>