Lift commands (#14)
* Create Lift Command

Tying the buttons to the lift. Not finished

* Finished lift command

Added the lift command using setDirection(), put in m_guitarHero and m_controller, and used axisGreaterThan and axisLessThant whileTrue using the guitar hero joystick

* Lift Commands

WIP, in the middle of breaking it up to handle the two different lift motors

* Lift Commands and Arm Functions

Finished the Lift Command to raise and lower the arms of the robot. Not sure if it works, we have to wait for Robot for testing :)